verify the signature
this should really be accurate with the signed algorithm. the message should be accurate, the signature should be accurate , the algorithm used to sign the signature should be the same, the public key should be exported and transferred to the other party safely and accurately. all this should be accurate to enable verification of the signature in python.